fix: error when trying to complete a resumed programme week
The completeWeek mutation was looking up the week to complete
based on the most recently created workout-week the user has.

If the user switches to another programme and then switched
back to the initial programme, the most recently created week
is _not_ their current week as displayed in the app.

This now takes into account the user's active programme so the
error should no longer occur.